adds header and add tokens
